From: Tony Finch <fanf@demon.net> Cc: libwww-perl@ics.uci.edu Subject: patch for libwww-perl-5.18 Message-Id: <E0xgTQM-0004Vz-00@fanf.noc.demon.net> Date: Fri, 12 Dec 1997 11:38:30 +0000
There's still a couple of bugs in LWP::Protocol::ftp that I submitted
a patch for a couple of weeks ago. The first is a lack of checking the
return value from Net::FTP::new; the second is using Net::FTP::message
in a list context when a scalar value is desired.

--------------------------

Also changed from "new HTTP::Response" to "HTTP::Response->new" syntax
because I prefer it.