From: BooK <book@netcourrier.com> Subject: Bug in HTML::Form 0.03 To: libwww@perl.org Date: Mon, 12 Mar 2001 00:57:05 +0100 (CET)
Hi,

I am new to this list, to which I subscribed because of the following:
the "action" method from HTML::Form has no effect when used to set the
value of the action. I am using HTML::Form 0.03, which I just downloaded
from CPAN.

Here's a patch to correct what seems to be a typo.