From: merlyn@stonehenge.com (Randal L. Schwartz) Subject: [patch] hacking HTTP::Daemon to handle CONNECT properly To: Gisle Aas <gisle@ActiveState.com> Cc: libwww@perl.org Date: 04 Mar 2001 14:07:07 -0800 Gnus-Warning: This is a duplicate of message <m1elwdgpdg.fsf_-_@halfdome.holdit.com>
HTTP::Daemon 1.22 ->get_request passes the URL-like string following
the method to make a relative based URL.  For "CONNECT" methods, this
is wrong, since no scheme is present and a hostname:port combo can
sometimes look like a scheme:hostname combo.

My patch is a bit of a hack, adding in the "telnet:" scheme when
CONNECT is used, but it seems to work nicely with the proxy server I'm
creating.

The non-upward compatibility warning is that you must now call
$request->url->host_port to get the host/port string for CONNECT,
rather than calling $request->url and using the entire string.  Since
I'm probably one of three people in the world doing this, I think it's
probably safe to just document that. :)