From: Liam Quinn <liam@htmlhelp.com> Subject: Patch for two WWW::RobotRules bugs To: <libwww@perl.org> Date: Fri, 20 Apr 2001 13:18:18 -0400 (EDT)
Hi,

Attached is a patch for two WWW::RobotRules bugs:

1. If I use <http://www.htmlhelp.com:80/robots.txt> as the robot_txt_uri,
WWW::RobotRules will not disallow access to
<http://www.htmlhelp.com/award/>, but it will disallow access to
<http://www.htmlhelp.com:80/award/>.  My patched version compares the host
and port instead of just the authority, so that both the /award/ URIs are
disallowed.

2. If a robots.txt has

User-agent: WDG_SiteValidator
Disallow: /foo

and my robot uses

User-Agent: WDG_SiteValidator/1.2.5

then /foo is not disallowed.  The substring comparison in the is_me method
is looking for "WDG_SiteValidator/1.2.5" within "WDG_SiteValidator"
instead of the other way around.