From: "Sean M. Burke" <sburke@cpan.org> Subject: no LWP::Protocol and LWP::Protocol::nogo To: gisle@activestate.com Date: Sat, 05 May 2001 00:42:32 -0600
I was grinding my teeth the other day about that fact that if one
created a CGI that said:

#!/bin/sh
echo Location: file:///etc/passwd
echo

and then directed a Perl-based HTTP proxy server to access that
URL, it would merrily follow redirection to that local file.

So I created a class, 'LWP::UserAgent::nogo', that one can name
as the implementor for any protocols (like 'file') that one wishes
to disable.