From: "Sean M. Burke" <sburke@cpan.org> Subject: fix for language-matching in HTTP::Negotiate To: gisle@ActiveState.com Date: Sun, 27 May 2001 18:41:49 -0600
This fixes some odd problems in HTTP::Negotiate's treatment of language
tags: first, it failed to smash case in one place, and second, it assumed
that if I accept language A, and you have language AB, then AB is fine.
It now says that if I accept A, and you have A-B that that's fine.

Would you believe I meant to make this fix /three/ years ago?