From: Norton Allen <allen@huarp.harvard.edu> Subject: [PATCH] Resolve "Short Write" To: libwww@perl.org Date: Tue, 7 Aug 2001 13:34:24 -0400 (edt)
The "short write" error reported by http.pm is a shorthand for
"Your operating system has exploited a perfectly legal feature
that apparently has not been encountered in other operating
systems to date." That feature is the possibility that syswrite
might return having only written part of the requested record,
and I have encountered it under QNX4 for file sizes in excess
of 10K bytes. This patch does the obvious loop to write out the
rest.

  -Norton Allen