From: Chris Nandor <pudge@pobox.com> Subject: Bundle::LWP problem (bad order) To: gisle@aas.no Date: Fri, 21 Sep 2001 18:59:34 -0400
Hi Gisle,

In Bundle::LWP, URI is first, and MIME::Base64 is third.  But MIME::Base64
is in URI's PREREQ_PM, so that things get out of whack.  The result is that
in a Bundle I am working on (which has a ton of things in it), URI etc.
doesn't get installed until the end, which throws other things off.
Temporarily, I put MIME::Base64 before Bundle::LWP in my Bundle, but
changing the order for your next release would probably be good.

Thanks,