From: "Sean M. Burke" <sburke@cpan.org> Subject: patch to WWW/RobotRules.pm To: Gisle Aas <gisle@aas.no> Date: Thu, 18 Jul 2002 04:54:03 -0600
This fixes a few doc typos, corrects a changed URL, and rephrases an
example or two.

Codewise, it fixes a few bugs:  it now restricts robots.txt-checking
to just http and https schemes; and it has new logic to more correctly
check the match of user-agents' name strings against the strings in
"User-agent:" lines in robots.txt files.