From: "Brian J. Murrell" <93d103a48bddc5503b63e92b382f5012@interlinx.bc.ca> Subject: libwww: no Content-Length in POST with no content fix To: gisle@ActiveState.com Date: Thu, 15 Nov 2001 11:14:09 -0500 Resent-From: "Brian J. Murrell" <93d103a48bddc5503b63e92b382f5012@interlinx.bc.ca>
Hello Gisle,

I ran across a small (bug?) annomly with libwww 5.60 and POSTing with
no content.  It seems that only if there is content in the POST
request that a Content-Length: header is generated.  Unfortunately,
Squid barfs at this.  It seems to want to know the content length even
if it is zero.

Here is a patch that seems to fix the problem -- at least here in my
application: