From: Ken Williams <ken@mathforum.org> Subject: Patch for LWP Makefile.PL To: libwww@perl.org Cc: gisle@ActiveState.com, macosx@perl.org Date: Fri, 11 Jul 2003 10:18:12 -0500
Hi,

There's still a flaw with the conflict-detection in LWP's Makefile.PL
that prevents /usr/bin/head from being overwritten by HEAD on Mac OS
X. The problem is that MakeMaker uses $Config{installscript} as the
installation location for EXE_FILES items, not $Config{sitebin} as was
assumed in the code.  Therefore the conflict never gets detected when
those two paths are different.

The patch below is against LWP 5.69.