From: Thomas Lotterer <thl@dev.de.cw.com> Subject: Re: OpenPKG patch for libwww To: Gisle Aas <gisle@ActiveState.com> Date: Thu, 16 Oct 2003 12:35:54 +0200 Reply-To: thl@dev.de.cw.com Organization: Cable & Wireless Deutschland GmbH, Munich, Germany.
On Thu, Oct 16, 2003, Gisle Aas wrote:

> Thomas Lotterer <thl@dev.de.cw.com> writes:
>
> > We noticed that you (or someone else) is currently working on the libwww
> > Perl module. Please note the the OpenPKG project applies a patch [1]
> > to this module which might be worth to be reviewed for taking into the
> > official distribution.
>
> Sure.  What problem does this actually fix? The patch does not really
> explain it.  Are you the author of the patch?
>

Hmm, yep, i am. Reviewing my own comment from the past shows some hard
facts to explain the solution by i missed to describe the problem. So do
my logs.

So i'm writing down what i remember in this paragraph. The code in
ftp.pm assumes that for a 550 file access error - and only for that
error - it hit a directory and changes the code path towards directory
listing. The problem was a Microsoft Windows based IIS FTP server
reporting a 450 file access error where both i and ftp.pm expected a 550
error. So the module tried a directory listing of a file. I wanted to
blame the FTP server for the issue so i reviewed the RFC but found out
that returing a 450 error was perfectly valid. It is my understanding
that in the DOS/Windows world writing a file by default explicitly
locks it. So during upload of a file additional trials to access exacly
that file will be temporarily blocked and IIS reports 450. In the UNIX
world the same situation would allow download of a partial (of course,
truncated) file.

The patch extends the decision of checking for a 550 return value to the
range 400 <= x < 600.

Maybe there is a better was to distinguish between files and directories
anyway, but implementing this would be a larger task.

I want to add that we at OpenPKG use a program called "vcheck" which
makes use of this module. It crawls the vendor download sites of over
600 packages bi-daily and you must assume that it hits almost any kind
of HTTP and FTP server that exists. And the patch did not make things
worse since we use it :-)