From: Liam Quinn <liam@htmlhelp.com> Subject: [PATCH] LWP::RobotUA case-sensitive check for "Disallow" To: libwww@perl.org Date: Sat, 3 Apr 2004 18:22:32 -0500 (EST)
LWP::RobotUA won't parse a robots.txt file if the file does not contain
"Disallow".  The check for "Disallow" is case sensitive, but according to
the robot exclusion standard, field names are case insensitive.  This
causes LWP::RobotUA to ignore some robots.txt files that it should
parse.

Attached is a patch that makes the check for "Disallow" case
insensitive. The patch is against libwww-perl 5.76 (RobotUA.pm 1.23).

--
Liam Quinn