From: "Michel Koppelaar" <Michel.Koppelaar@kb.nl> Subject: media.types To: <gisle@aas.no> Date: Tue, 17 Feb 2004 14:58:00 +0100
Hi,

I was playing with the function guess_media_type() in LWP, when I found
that it would report a .mov file to be of type video/x-sgi-movie,
whereas nowadays it's more common to use that extension for quicktime
movies.

Shouldn't the line
video/x-sgi-movie               mov
in the file media.types be removed and the quicktime line changed to
video/quicktime               qt mov?

Thanks,

Michel Koppelaar
National Library of the Netherlands