HTML::Form Module and <button> element clicks
<button> element is not supported correctly in HTML::Form module.  Button
elements are not clickable using the method ->click().  They are not even
parsed into an input object due to its type being set to "IgnoreInput" in the
%type2class hash.

The default action for <button> should be equivalent to <input type="submit">,
at least according to HTML 4.01 and XHTML 1.0 strict DTDs:
http://www.w3.org/TR/html401/interact/forms.html#edef-BUTTON
http://www.w3.org/TR/xhtml1/dtds.html#dtdentry_xhtml1-strict.dtd_button

I could understand ignoring the button for other button types such as
type="reset" or type="button", but not "submit" which is default.  Is there any
particular reason why button elements are not supported?