Disabled, checked radiobutton being submitted [RT#33512]
When I have a disabled, checked radiobutton and a hidden field, both
with the same name, the disabled radiobutton "succeeds" and it's value
is submitted by HTML::Form.

If I understand section 17.12.1 "Disabled controls"¹ in the HTML 4.01
standard correctly, the radiobutton should not succeed when disabled,
and this following test should be valid.

¹ <http://www.w3.org/TR/html4/interact/forms.html#adef-disabled>