Small documentation issues [RT#31346]
From: http://tincho.myopenid.com/

There are a couple of documentation errors in lwp-request and
Cookies.pm, reported in the Debian BTS, which we fixed and now we want
to send them to you so you can also apply them to your code. I hope you
find it useful. Below is the detailed description (the bugs can be read
in http://bugs.debian.org/bugnr):

- bin/lwp-request: mention GET, POST and HEAD in the POD, fixed the
  discrepancies between the synopsis, the POD description and the
  cmdline help (-C and -P options, plus short options summary), also added some
  standard markup in POD synopsis (Closes: #396968, #408426, #354905)

- libwww-perl/lib/HTTP/Cookies.pm: updated Netscape cookies spec's URL
  (Closes: #422732)