Make lwp-request honour option -b while dumping links [RT#31347]
From: http://tincho.myopenid.com/

This is a bug that was reported to Debian some time ago, and we solved
it with a small patch (included). This is the original bug text:

suppose you have retrieved a copy of http:/a.b.c/d/e.html and saved it
in a local repository /my/repo, and then you want to extract links
(based on the original url) from the saved file.
the first thing that comes to my mind,
$ lwp-request -o links -b http:/a.b.c/d/e.html file:/my/repo/e.html
does not work, because lwp-request picks the base url from the url it is
parsing (namely file:/my/repo/e.html)