Don't allow HTTP::Message content to be set to Unicode strings.
A common error is that Unicode strings sneak into HTTP::Request
objects and confuse the protocol modules which assumes bytes.
This change try to catch this problem early.  It's still possible
to insert chars when you manipulate the $mess->content_ref directly
but there is no efficient way to police that, so we don't.