Allow malformed chars in $mess->decoded_content [RT#17368]
If a webserver returns badly encoded data (bad utf8), I still need to be
able to get it decoded. Please add an option to ignore problems, e.g
like with the attached patch.

IMHO, since that's the default in Encode, ignoring a bad encoding might
be the better default. Also, maybe a way for setting a global default
error mode in Encode would be the better solution. However, I don't
think that's for me to decide and I'd be happy if my patch makes it into
the distribution.

Thanks, Rolf.