Allow 100, 204, 304 responses to have content [RT#17907]
IIS v5.0 appears to violate RFC 2616 Section 4.3 by sending content in
its 403 (Not Modified) HTTP 1.1 responses. Here is an example found in
a production server on the web:

  --BEGIN--
  HTTP/1.1 304 Not Modified
  Server: Microsoft-IIS/5.0
  Date: Tue, 28 Feb 2006 23:45:12 GMT
  Pragma: No-cache
  Cache-Control: Private
  P3P: CP="ALL DSP CURa IVAa HISa OUR NOR IND UNI NAV INT STA PRE"
  Expires: -1
  ETag: "9e496ab930c5c51:e1a"
  Connection: close
  Transfer-Encoding: chunked

  0

  --END--

This is a problem for Net::HTTP because when it sees 304 responses it
does not read any content and, leaving the content data on the socket,
will mis-parse the next response on that connection.

This section of the RFC says that content is not allowed on 304, 204, or
100 responses, so this response is in violation of the RFC, but it only
specifically mentions that responses to HEAD requests may have
Content-Length and Transfer-Encoding headers that must be ignored.