Add strict mode to HTML::Form
This makes HTML::Form non-strict by default and if you want it to
be strict you need to request that explictly.  When you turn on
strict it's even stricter than it used to be as it will now croak
if you try to assign to readonly fields or if you assign values
that are longer than 'maxlength'.