Make split_header_words() lower case returned tokens/keys.
Found many places that were buggy because it just assumed this to
be the case.  The only exception I have found is when parsing cookies
where we want to preserve the case as specified from the server.
Since we leave the old behaviour as _split_header_words() it should
be trivial to adopt code that happened to depend on the old behaviour.