HTTP::Cookies::Microsoft now handles user names with space [RT#40917]
I use $VERSION = "5.810" of Cookies::Microsoft in ActivePerl 5.8, but
find sometimes the cookie added to request is not complete:
for example, when the target is books.google.cn, only books.google.cn
cookie is added, but not google.cn.

I tracked down to a line reading cookie files from index.dat

my user name is "shaohua li", but the browser stores google.cn's
cookie in "shaohua_li@google[5].txt", as filemon reveals.

So here is a mismatch and this entry is skipped.