Fixup test failure with perl-5.8.8 and older; qr/$/m doesn't work
The new way in which some of the LWP tests are expressed in version 5.824
tickles a bug in perl that I discovered about five years ago (#27028) and
which is present in all versions up to 5.8.8.  The bug is that a qr/$/m
effectively has the /m ignored.  This is causing false test failures.
Fortunately, there's an easy workaround (attached).

http://rt.perl.org/rt3/Public/Bug/Display.html?id=27028