new tests for max_size and 206 responses [RT#46230]
Attached is a patch against the current "git" state that adds some more
tests for "max_size" and 206 http responses (all the the tests pass).

I was also curious to know if "206" would be returned in the case that
max_size was used, the max_size requested was greater than the size of
the document. But, that behavior is provided by the server, not
LWP::UserAgent, so we don't have control over it.