mirror should die in case X-Died is set [RT#48236]
In some situations LWP::UserAgent::mirror gives a bad error message in
the form "Transfer truncated: only out of ... bytes received". This
seems to happen if something calls die() while in the request callback.
This only sets X-Died, but nothing checks for this value. My attached
patch changes this to check X-Died, and if it is set, to die with that
value, which gives better error messages.