Improve diagnostics from LWP::UserAgent::mirror [RT#48869]
A similar issue is described in #48236. I happened to come across the
same problem and found a different spot where the diagnostics can be
improved. Because the user had no write permissions to the directory the
tmpfile was not even written and the stat() call was unsuccessful. I
suspect this has a corollary in some unsuccessful close() further up the
chain.

The patch seems orthogonal to Slaven's patch in 48236, so I make this a
new RT ticket.