Improved handling of the User-Agent header [RT#48461]
If the empty string is passed as the 'agent' parameter to the
constructor, then no User-Agent will be announced in the requests.
This worked before when $ua->agent('') was called explictly, but
not for the value passed to the constructor.

If the 'default_header' passed to the constructor already has the
'User-Agent' header set then don't override this will the default
libwww-perl value.

If both the 'default_header' and the 'agent' is passed then the value
passed in 'agent' wins.