Remove needless (and actually harmful) local $_
I came across a bug in HTTP::Message::content_charset. It has a
'local $_', which is unnecessary, since foreach loops already
localise their topic. In fact, they do it in a safer way that is more
like local *_=\do{my$x}. Using local $_ causes problems if $_ is tied.
The attached patch removes the offending line and adds a test for it.
This did actually occur in real code, and is not just a theoretical
problem.