Strip out empty lines separated by CRLF
Before the patch, an embedded CRLF would be allowed through with a
space added after it. This is spec-compliant but appears not to be
the intent, based on the regex to remove empty lines in headers.  If
a client accepted "CRLF\sCRLF\s" as a valid header end, then a CRLF
injection attack would have been possible.