Make the authentication of multiple realms work.
There is a 'feature' to stop the too many requests
being sent that checks if the auth password has changed when trying
auth.  In the case of us having multple realms we try the same
user/password combo, just a different realm.

This is a naive commenting out of that check which makes all the tests
pass, but it also means there are potentially more requests being sent.
I need to figure out how to deal with our use case so it works without
completely clobbering this check.  Of course it would be good if there
was a test that failed with this change to prove that the original
feature was working too.