Changes *should* have content in PRs.
Hopefully these comments will show that it is very much intentional,
hence why the file is doctored here so that the test is always run on
travis, but failures are acceptable on the master branch (where smaller
changes may have been pushed that do not require a Changes entry).
However, every pull request should have a user-oriented description of
the change.