Allow IPv6 proxy hosts starting with [.
Altered the regular expression that checked for a sane-ish looking
URL for our proxy host:

 if $url =~ /^https?:/ && $url !~ m,^https?://\w,;

Given this expression, a proxy host in the form of
http://[::1]:33485/ would fail because it starts with a left bracket.

The expression above has been altered to:

 if $url =~ /^https?:/ && $url !~ m,^https?://[\w[],;

Which will allow it to begin with a word character or a left bracket.

This will resolve issue #237

Also, edit the Travis config to ensure we test against threaded Perls.