Fix threaded Perl tests.
Some threaded versions of Perl hit an odd failure on the
xt/author/live/jigsaw/auth-d.t tests.

Failing back to an empty array ref when there's nothing left to grab
from the test set masks the problem. What we haven't yet had time to
investigate is _why_ the tests are asking for more data than what we've
provided. There's some odd quirk going on that we should investigate and
figure out when time permits. THIS IS A BANDAID.