Add the patch method to LWP::UserAgent.
Much like the previously added `put` method, the `patch` method will
require a specific version of HTTP::Request::Common. This time, though,
we will need at least version 6.12.

This is largely a copy-paste of the already added `put` method. All of
the current tests we have in place that test the `put` method were
duplicated for the new `patch` method. The same logic gate that only
tests when we have the proper version of HTTP::Request::Common was added
for the 6.12 check.

This makes LWP::UserAgent a bit simpler to work with RESTful endpoints
in that we now directly support the most common HTTP verbs: `HEAD`,
`GET`, `DELETE`, `PUT`, `POST`, `PATCH`.

Fixes #116