Fix a looping regression for 303 responses without a Location header
The change in 6.45 had a side effect of making it look at max_redirect
only when the Location header is set.

The libapache2-mod-perl2 test suite implements a test server that answers
with HTTP code 303 without including such a header.

RFC 7231 is not very clear on whether Location is required, although
the response doesn't make much sense without it.

Some googling reveals this is not a new question.

  https://stackoverflow.com/questions/16194988/for-which-3xx-http-codes-is-the-location-header-mandatory

For the sake of robustness I suppose libwww-perl needs to be fixed.

Bug-Debian: https://bugs.debian.org/962904