Remove block of code which creates message-digest auth request field
Fixes #368

To quote Florian Lohoff in
https://github.com/libwww-perl/libwww-perl/issues/368:

"I found the corresponding RFCs or better - Drafts which mentioned this
type of authentication - I attached that to the Debian Bug - Basically
this authentication type was once mentioned in a Draft and was dropped
from the next draft. Implementation dates back to inbetween these 2
drafts. Somebody forgot to remove the code when this was dropped in the
standardisation process."

See also https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=974572