Replace remaining uses of "base" with "parent"
Base says "Unless you are using the fields pragma, consider this module
discouraged in favor of the lighter-weight parent.". We aren't using
fields so let's follow their advice, parent is also core in 5.10+.

We were already using parent in one place (Protocol::http) though it was
undeclared in our deps so this fixes that. Depend on the minimum version
to support "-norequire".